+++ wgpu-hal/src/gles/wgl.rs
@@ -885,6 +885,7 @@ impl crate::Surface for Surface {
             inner: super::TextureInner::Renderbuffer {
                 raw: sc.renderbuffer,
             },
+            externally_owned: false,
             drop_guard: None,
             array_layer_count: 1,
             mip_level_count: 1,